lint: Fix pylint warning about default `display_value`
pylint warns "No value for argument 'cstruct' in lambda call" because
`serialize()` calls the `display_value` function with two arguments.
But if we define it as `lambda field, cstruct: cstruct`, then we have
an error:

    <lambda>() takes 2 positional arguments but 3 were given

This is because the lambda is bound to the instance and Python passes
the instance as the first argument. So we end up with 3 arguments.

I tried to shut up pylint by ignoring "no-value-for-parameter" for
this line, but then pylint warns about its "useless suppression".