setup: Don't require pyramid_chameleon and pull Pyramid
We don't need Pyramid, we only need Chameleon (which is already a
requirement of Deform).

However, we would like to test the `includeme` function and it's
better if we do it with a real `pyramid.config.Configurator` object.
This means that we need to install Pyramid for tests, even though it's
quite a big requirement that itself pulls many things. We could mock
this `Configurator` object but we'll need Pyramid anyway for the
upcoming browser (Selenium) tests.